Revive tests
